move errors from app-globals to its own service
* Fixes #339
