+++ src/abstract-list-field/abstract-list-field.component.ts
@@ -23,7 +23,7 @@ import { ChangeDetectorRef } from '@angular/core';
 import { List } from 'immutable';
 
 import { AbstractFieldComponent } from '../abstract-field';
-import { JsonStoreService, AppGlobalsService, PathUtilService } from '../shared/services';
+import { JsonStoreService, AppGlobalsService, PathUtilService, ErrorsService } from '../shared/services';
 import { JSONSchema, JsonPatch } from '../shared/interfaces';
 
 /**
@@ -39,10 +39,11 @@ export abstract class AbstractListFieldComponent extends AbstractFieldComponent 
   path: Array<any>;
 
   constructor(public appGlobalsService: AppGlobalsService,
+    public errorsService: ErrorsService,
     public jsonStoreService: JsonStoreService,
     public pathUtilService: PathUtilService,
     public changeDetectorRef: ChangeDetectorRef) {
-    super(appGlobalsService, pathUtilService, changeDetectorRef, jsonStoreService);
+    super(appGlobalsService, errorsService, pathUtilService, changeDetectorRef, jsonStoreService);
   }
 
   /**
