+++ src/primitive-list-field/primitive-list-field.component.ts
@@ -24,7 +24,7 @@ import { Component, Input, ChangeDetectionStrategy, ChangeDetectorRef } from '@a
 import { List } from 'immutable';
 
 import { AbstractListFieldComponent } from '../abstract-list-field';
-import { AppGlobalsService, JsonStoreService, PathUtilService } from '../shared/services';
+import { AppGlobalsService, JsonStoreService, PathUtilService, ErrorsService } from '../shared/services';
 import { JSONSchema } from '../shared/interfaces';
 
 @Component({
@@ -41,10 +41,11 @@ export class PrimitiveListFieldComponent extends AbstractListFieldComponent {
   @Input() path: Array<any>;
 
   constructor(public appGlobalsService: AppGlobalsService,
+    public errorsService: ErrorsService,
     public jsonStoreService: JsonStoreService,
     public pathUtilService: PathUtilService,
     public changeDetectorRef: ChangeDetectorRef) {
-    super(appGlobalsService, jsonStoreService, pathUtilService, changeDetectorRef);
+    super(appGlobalsService, errorsService, jsonStoreService, pathUtilService, changeDetectorRef);
   }
 
 }
